# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Elham Multiplast CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List all contacts/leads
      parameters:
        - name: salesOwnerId
          in: query
          schema:
            type: integer
        - name: city
          in: query
          schema:
            type: string
        - name: unit
          in: query
          schema:
            type: string
        - name: industry
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a new contact/lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    patch:
      operationId: updateContact
      tags: [contacts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List activities (with optional filters)
      parameters:
        - name: dealId
          in: query
          schema:
            type: integer
        - name: contactId
          in: query
          schema:
            type: integer
        - name: userId
          in: query
          schema:
            type: integer
        - name: upcoming
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Log an activity or schedule a follow-up
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
  /activities/{id}:
    patch:
      operationId: updateActivity
      tags: [activities]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
    delete:
      operationId: deleteActivity
      tags: [activities]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, username, role, colorCode, unit]
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [admin, sales]
        colorCode:
          type: string
        unit:
          type: string
          enum: [Himatnagar, Surat, Rajkot, All]
        createdAt:
          type: string
...
    Contact:
      type: object
      required: [id, name, mobile, salesOwnerId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        mobile:
          type: string
        email:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        salesOwnerId:
          type: integer
        salesOwner:
          $ref: "#/components/schemas/User"
        otherPhone:
          type: ["string", "null"]
        otherEmail:
          type: ["string", "null"]
        leadSource:
          type: ["string", "null"]
          enum: [IndiaMart, TradeIndia, "Social Media", Organic, Email, Other, null]
        city:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
          enum: [Himatnagar, Surat, Rajkot, null]
        industry:
          type: ["string", "null"]
          enum: ["Liquid Detergent", Lubricant, "Agro Chemical & Pesticide", "Edible Oil", Veterinary, Other, null]
        tags:
          type: ["string", "null"]
          enum: [Interested, "Category B", "Category C", null]
        inquiryDate:
          type: ["string", "null"]
        lastCallDate:
          type: ["string", "null"]
        nextCallDate:
          type: ["string", "null"]
        createdAt:
          type: string
    ContactInput:
      type: object
      required: [name, mobile, salesOwnerId]
      properties:
        name:
          type: string
        mobile:
          type: string
        email:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        salesOwnerId:
          type: integer
        otherPhone:
          type: ["string", "null"]
        otherEmail:
          type: ["string", "null"]
        leadSource:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        inquiryDate:
          type: ["string", "null"]
        lastCallDate:
          type: ["string", "null"]
        nextCallDate:
          type: ["string", "null"]
    ContactUpdate:
      type: object
      properties:
        name:
          type: string
        mobile:
          type: string
        email:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        salesOwnerId:
          type: integer
        otherPhone:
          type: ["string", "null"]
        otherEmail:
          type: ["string", "null"]
        leadSource:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        tags:
          type: ["string", "null"]
        inquiryDate:
          type: ["string", "null"]
        lastCallDate:
          type: ["string", "null"]
        nextCallDate:
          type: ["string", "null"]
...
    Activity:
      type: object
      required: [id, dealId, type, createdAt]
      properties:
        id:
          type: integer
        dealId:
          type: integer
        contactId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [Call, WhatsApp, Email, Note, FollowUp]
        notes:
          type: ["string", "null"]
        followUpDate:
          type: ["string", "null"]
        followUpType:
          type: ["string", "null"]
          enum: [Call, WhatsApp, Email, null]
        createdBy:
          type: ["integer", "null"]
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
    ActivityInput:
      type: object
      required: [dealId, type]
      properties:
        dealId:
          type: integer
        contactId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [Call, WhatsApp, Email, Note, FollowUp]
        notes:
          type: ["string", "null"]
        followUpDate:
          type: ["string", "null"]
        followUpType:
          type: ["string", "null"]
    ActivityUpdate:
      type: object
      properties:
        notes:
          type: ["string", "null"]
        followUpDate:
          type: ["string", "null"]
        followUpType:
          type: ["string", "null"]
...