# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Elham Multiplast CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  name: string;
  mobile: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  companyName?: string | null;
  salesOwnerId: number;
  /** @nullable */
  otherPhone?: string | null;
  /** @nullable */
  otherEmail?: string | null;
  /** @nullable */
  leadSource?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  tags?: string | null;
  /** @nullable */
  inquiryDate?: string | null;
  /** @nullable */
  lastCallDate?: string | null;
  /** @nullable */
  nextCallDate?: string | null;
}