# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Elham Multiplast CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';
import type { UserUnit } from './userUnit';

export interface User {
  id: number;
  name: string;
  username: string;
  role: UserRole;
  colorCode: string;
  unit: UserUnit;
  createdAt?: string;
}